Looks like workers implementation works Done static File::Open(..) Done File::Close(..)
